requirements.txt
Pillow==7.0.0
progress==1.5
